Modify equals button event listener and operate function
The equals button event listener was modified to check if answer is a
floating point number before it is rounded in case answer is a string
or so no unneccessary rounding takes place. The operate function was
modified so that the decimal point "." is treated as zero if that is
the only value the user inputs for num1/num2.